+++ emqx-http-sdk/src/main/java/io/github/gcdd1993/emqx/sdk/http/model/request/SubscribeRequest.java
@@ -26,7 +26,7 @@ public class SubscribeRequest {
     /**
      * QoS 等级
      */
-    private Integer qos;
+    private Long qos;
 
     /**
      * 客户端标识符
