+++ emqx-http-sdk/src/main/java/io/github/gcdd1993/emqx/sdk/http/model/response/AlarmDetailsDto.java
@@ -15,5 +15,5 @@ import lombok.NoArgsConstructor;
 public class AlarmDetailsDto {
 
     @JsonProperty("high_watermark")
-    private Integer highWatermark;
+    private Long highWatermark;
 }
